Fix a theme color and combine the hooks for tracking transfers
